// apps/journeys-admin/__generated__/AiCreateCloudflareUploadByFileMutation.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: AiCreateCloudflareUploadByFileMutation
// ====================================================

export interface AiCreateCloudflareUploadByFileMutation_createCloudflareUploadByFile {
  __typename: "CloudflareImage";
  id: string;
  uploadUrl: string | null;
}

export interface AiCreateCloudflareUploadByFileMutation {
  createCloudflareUploadByFile: AiCreateCloudflareUploadByFileMutation_createCloudflareUploadByFile;
}